resolve import cycles
