+++ mex/common/sinks/registry.py
@@ -1,13 +1,49 @@
-from typing import TYPE_CHECKING, Final
+from collections.abc import Generator, Iterable
+from itertools import tee
+from typing import Final
 
+from mex.common.models import AnyExtractedModel
+from mex.common.settings import BaseSettings
 from mex.common.sinks.backend_api import BackendApiSink
+from mex.common.sinks.base import BaseSink
 from mex.common.sinks.ndjson import NdjsonSink
-from mex.common.types import Sink
-
-if TYPE_CHECKING:
-    from mex.common.sinks.base import BaseSink
+from mex.common.types import Identifier, Sink
 
 _SINK_REGISTRY: Final[dict[Sink, type["BaseSink"]]] = {}
+
+
+class _MultiSink(BaseSink):
+    """Sink to load models to multiple sinks simultaneously."""
+
+    # This class is private because it should only be acquired by calling `get_sink`.
+
+    _sinks: list[BaseSink] = []
+
+    def __init__(self) -> None:
+        """Instantiate the multi sink singleton."""
+        settings = BaseSettings.get()
+        for sink in settings.sink:
+            if sink in _SINK_REGISTRY:
+                sink_cls = _SINK_REGISTRY[sink]
+                self._sinks.append(sink_cls.get())
+            else:
+                msg = f"Sink function not implemented: {sink}"
+                raise RuntimeError(msg)
+
+    def close(self) -> None:
+        """Close all underlying sinks."""
+        for sink in self._sinks:
+            sink.close()
+
+    def load(
+        self,
+        models: Iterable[AnyExtractedModel],
+    ) -> Generator[Identifier, None, None]:
+        """Load models to multiple sinks simultaneously."""
+        for sink, model_gen in zip(
+            self._sinks, tee(models, len(self._sinks)), strict=False
+        ):
+            yield from sink.load(model_gen)
 
 
 def register_sink(key: Sink, sink_cls: type["BaseSink"]) -> None:
@@ -35,10 +71,7 @@ def get_sink() -> "BaseSink":
     Returns:
         A function that pours the models into all configured sinks
     """
-    # break import cycle, sigh
-    from mex.common.sinks.base import MultiSink
-
-    return MultiSink.get()
+    return _MultiSink.get()
 
 
 # register the default providers shipped with mex-common
